Rename parser to scanner module to better reflect its functionality
