+++ src/core/mod.rs
@@ -1,1 +1,1 @@
-pub mod parser;
+pub mod scanner;
