+++ src/core/scanner/mod.rs
